backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
spotipy==2.23.0
python-dotenv==1.0.0
pydantic==2.5.0
aiosqlite==0.19.0python